[dependencies] - Move peters-toolbelt to a devDependency.
Because `peters-toolbelt` is a dependency, in `npm v3`, it brings its
dependencies to the root `node_modules` directory. This causes
collisions with webpack config referencing `babel` instead of
`babel_loader` when using a nested directory structure. I believe all
of the dependencies of `peters-toolbelt` are devDependencies except for
`react` which we have already specified as a dependency here. This
should be harmless and fix issues upstream in `npm v3`. Thanks!